Split off serializer_core to match deserializer_core
