+++ rbx_xml/src/lib.rs
@@ -13,6 +13,7 @@ mod deserializer;
 mod deserializer_core;
 mod error;
 mod serializer;
+mod serializer_core;
 mod types;
 
 #[cfg(test)]
